[DAT-8952] Fixes antivirus updates failing to download
This job writes to /tmp/ by default the only writeable location in Lambda
If AWS fails to provision a new instance for repeated runs can fill /tmp/
as it's not cleared between runs if the instance is re-used

This commit clears /tmp/ after a run. It also updates the Dockerfile with
several dependencies as the base image has changed since our last
rebuild.

Tested in production becaue things were broke.
<insert bill oreilly gif here>